Move material that doesn't need auditing out of number_specs.rs
